Renderer toolkit: "RendererToolkit::CacheManager::needsToBeCompiled()": Added format version so we can detect file format version changes and enforce compiling even if the source data has not been changed. With this the cache hopefully doesn't break so often anymore during development which changes things, of course.

+++ Renderer/RendererRuntime/include/RendererRuntime/Resource/MaterialBlueprint/Loader/MaterialBlueprintFileFormat.h
@@ -63,7 +63,7 @@ namespace RendererRuntime
 			struct Header
 			{
 				uint32_t formatType;
-				uint16_t formatVersion;
+				uint32_t formatVersion;
 				uint32_t numberOfProperties;
 				uint32_t numberOfShaderCombinationProperties;
 				uint32_t numberOfIntegerShaderCombinationProperties;
